Add zlib-ng to Meson as a default-enabled zlib replacement
zlib-ng is an actively maintained and modern replacement for zlib.
It includles SIMD optimizations for:
- Intel and AMD's SSE2/3/4, AVX (and more)
- ARM's NEON,
- Power ISA's Altivec, and
- RISC-V's "V" Vector (RVV) extensions

It's enabled by default for anyone building from source, provided
they have both CMake and Meson version 1.3.0 (as this version
of Meson added the cmake subproject handling for wraps).

If those aren't provided or available, then the existing zlib depedency
and wrap will provide the deflate functionality.

Status is logged in the meson setup summary list:

  Build Summary
    zlib provider: built-in

  Subprojects
    zlib-ng: YES